Use rocksdb to cache vars `from_command` output
This commit introduces a cache based on rocksdb to store
the output of `from_command` vars.

It is usefull in the case were you have vars that make call's to remote
api's that are slow.

One example of this would be listing the ip-address and instances names
of EC2 machines.